+++ build.gradle.kts
@@ -8,10 +8,11 @@ buildscript {
     }
     dependencies {
         classpath("com.android.tools.build:gradle:${Deps.Version.AndroidGradle}")
+        classpath("com.google.gms:google-services:4.3.10")
         classpath("org.jetbrains.kotlin:kotlin-gradle-plugin:${Deps.Version.Kotlin}")
         classpath("com.google.dagger:hilt-android-gradle-plugin:${Deps.Version.Hilt}")
-        classpath("com.google.firebase:firebase-appdistribution-gradle:4.0.0")
-        classpath("com.google.firebase:firebase-crashlytics-gradle:2.9.8")
+        classpath("com.google.firebase:firebase-appdistribution-gradle:3.0.0")
+        classpath("com.google.firebase:firebase-crashlytics-gradle:2.8.1")
     }
 }
